requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pytesseract
pdfminer.six
pillow
chromadb
unstructured
sentence-transformers
pdf2image
unstructured-inference
pi-heif
langchain